Fix GL context lost when app is put in background
Just a quick update to synchronize this branch with master.
Also this is now fully working on SGS3.

In order to restore the GL context the "SDL_GL_CreateContextGL" function
is called after a "nativeResume". The "createEGLContext" function is
then called on the java side.

I removed the SDL_ANDROID_BLOCK_ON_PAUSE define as we already block the
core via "emuPause".